refactor: Notification 타입 분리
