+++ src/fsd_entities/notification/config/types.ts
@@ -1,0 +1,8 @@
+export interface Notification {
+  targetId: string;
+  title: string;
+  noticeType: string;
+  body: string;
+  isRead: boolean;
+  notificationLogId: string;
+}
